image-recognition/requirements.txt
boto3==1.35.71
Flask==3.1.0
flask-swagger-ui==4.11.1
fuzzywuzzy==0.18.0
pandas==2.2.3
pillow==11.0.0
pip-chill==1.0.3
python-Levenshtein==0.26.1
requests==2.32.3
zxing==1.0.3
gunicorn==21.2.0
azure-cognitiveservices-vision-computervision==0.9.1
